+++ NEWS.md
@@ -54,3 +54,4 @@ Pre-1.0.0 numbering scheme: 0.x will indicate releases, while 0.0.x will indicat
   steps has changed the appropriate values
 - produce length 0 `dist_quantiles()`
 - add functionality to calculate weighted interval scores for `dist_quantiles()`
+- Add `step_epi_slide` to produce generic sliding computations over an `epi_df`